Feat: print out info for StoreField in HIR level
This commit marks the beginning of transitioning our escape-info-checker
from LIR to HIR.

The motivation for this change is that method names are lost at the LIR
level if methods are inlined during HIR level, which results in insufficient
information being avaialble for our checker.

By moving the checker to HIR level, we can get sufficient info.