Feat: move checker from LIR to HIR
Refactor the checker to determine the need for MFENCE while inserting
the StoreField class into HIR graph.

To propagate this information to the LIR level, a boolean field has been
added to the StoreField class.

As a result, at the LIR level, it's now sufficient to simply check the
boolean field.